move main ingest logic to core lib + mdb specific stuff to our implementation
