+++ packages/ingest-mongodb-public/src/sources/snooty/SnootyDataSource.ts
@@ -1,15 +1,15 @@
 import { createInterface } from "readline";
 import fetch from "node-fetch";
-import { DataSource } from "../DataSource";
+import { DataSource } from "mongodb-rag-core/ingest";
 import { snootyAstToMd, getTitleFromSnootyAst } from "./snootyAstToMd";
-import { ProjectBase } from "../ProjectBase";
+import { ProjectBase } from "mongodb-rag-core/ingest";
 import {
   getTitleFromSnootyOpenApiSpecAst,
   snootyAstToOpenApiSpec,
 } from "./snootyAstToOpenApiSpec";
-import { Page } from "../../../Page";
-import { PageFormat } from "../../../PageFormat";
-import { logger } from "../../../services/logger";
+import { Page } from "mongodb-rag-core";
+import { PageFormat } from "mongodb-rag-core";
+import { logger } from "mongodb-rag-core";
 
 // These types are what's in the snooty manifest jsonl file.
 export type SnootyManifestEntry = {
@@ -28,10 +28,10 @@ export type SnootyPageEntry = SnootyManifestEntry & {
 /**
   Represents metadata in a Snooty manifest file.
  */
-  export type SnootyMetadataEntry = SnootyManifestEntry & {
-    type: "metadata";
-    data: {title?: string};
-  };
+export type SnootyMetadataEntry = SnootyManifestEntry & {
+  type: "metadata";
+  data: { title?: string };
+};
 
 /**
   A node in the Snooty AST.
@@ -65,9 +65,9 @@ export type SnootyPageData = {
 /**
   A Snooty Data API metadata object. This contains project-level information, such as the site name.
  */
-  export type SnootyMetadata = {
-    title?: string;
-  };
+export type SnootyMetadata = {
+  title?: string;
+};
 
 export type SnootyProjectConfig = ProjectBase & {
   type: "snooty";
@@ -168,7 +168,7 @@ export const makeSnootyDataSource = ({
       const stream = createInterface(body);
       const linePromises: Promise<void>[] = [];
       const pages: Page[] = [];
-      let siteTitle: string | undefined = undefined
+      let siteTitle: string | undefined = undefined;
       await new Promise<void>((resolve, reject) => {
         stream.on("line", async (line) => {
           const entry = JSON.parse(line) as SnootyManifestEntry;
@@ -231,11 +231,11 @@ export const makeSnootyDataSource = ({
       });
       await Promise.allSettled(linePromises);
       // add metadata to all the pages
-      for(const page of pages){
+      for (const page of pages) {
         if (!page.metadata) {
-          page.metadata = {}
+          page.metadata = {};
         }
-        page.metadata.siteTitle = siteTitle 
+        page.metadata.siteTitle = siteTitle;
       }
       return pages;
     },
