+++ packages/ingest-mongodb-public/src/sources/snooty/snootyAstToOpenApiSpec.ts
