+++ packages/mongodb-rag-core/package.json
@@ -24,13 +24,13 @@
     "build",
     "README.md"
   ],
+  "main": "./build/index.js",
   "exports": {
     ".": "./build/index.js",
     "./ingest": "./build/ingest/index.js",
     "./ingest/embed": "./build/embed/index.js",
     "./ingest/pages": "./build/pages/index.js",
-    "./ingest/sources": "./build/sources/index.js",
-    "./ingest/sources/snooty": "./build/sources/snooty/index.js"
+    "./ingest/sources": "./build/sources/index.js"
   },
   "scripts": {
     "clean": "rm -rf build",
