+++ packages/mongodb-rag-core/src/ingest/sources/index.ts
@@ -1,14 +1,11 @@
 export * from "./AcquitRequireMdOnGithubDataSource";
 export * from "./DataSource";
-export * from "./DevCenterDataSource";
+export * from "./extractMarkdownH1";
 export * from "./GitDataSource";
 export * from "./GitHubDataSource";
 export * from "./MdOnGithubDataSource";
 export * from "./ProjectBase";
-export * from "./RstOnGitHubDataSource";
 export * from "./handleHtmlDocument";
 export * from "./removeMarkdownImagesAndLinks";
-export * from "./snooty";
-export * from "./mongodb-university";
 export * from "./LangchainDocumentLoaderDataSource";
 export * from "./CodeOnGithubTextDataSource";
